parser/src/requirements.txt
python-dotenv==1.0.1
selenium==4.27.1
huggingface_hub==0.28.0
requests==2.32.3
pytz==2025.1
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6